add a configurable procedure sample
This class is meant to be used for cases where we'd like to prepare a
bunch of procedures to be ran at some later point. We can add and remove
lambdas (by tags) to a ConfigurableProcedure object, and then invoke the
entire thing once ready.

A class can hypothetically hold such an instance and simply add/remove
lambdas it were supposed to execute as user selects/unselects options.