add when-is replacemnt based on reflection
In this sample I use reflection in order to obtain access to private
fields of the BusinessLogic class from an api response subclass.

While this both preserves encapsulation of the BusinessClass's private
logic and allows to use polymorphism instead of when-is (instanceof
checks essentially) blocks, it doesn't really feel like it's solving the
problem. Predominantly, the privacy issue has been tackled in the most
brutal way I can think of currently. Yes, the functions are private,
but...

Helper functions could be added to simplyfy the use of reflection
on the BusinessLogic class.