add android vector to SVG converter
Since android uses a slightly different format, than SVG, for vector
graphics, it's sometimes helpful to be able to turn graphics converted
into it back to their SVG form Android studio had imported them
from.

Note that this is not a performance champion, it's just right for single
cases sporadically and should be perfectly fine dealing with such a
load. For multi-million arrays of xml drawables, however, consider
something else.