introduce vetoable sample
A vetoable delegate allows for neat mechanism of rejecting values which
fail to meet our arbitrary criteria.

While cute, it might be less than clean when reading code full of those,
where a property might just reject a value whereas we assume it did
assume it and continue our debugging with such a thought.