add request runner samples
Two competing solutions were added - one focused on simplicity, the
other on extensibility and clarity. The "convenientRunner" sample is the
latter, and seems superior. Its "Convenience" class, provided within the
"body" scope of the run function, is not only capable of providing
useful pieces of data known today, but can trivially be modified to
encompass more data as needed. In addidion to this, it can also include
helper functions etc.

Similar approach could be applied to the onException and onSuccess
methods, but it's less crucial there, if needed at all.